psp-10275 edit properties name and display error when property in file activity.

+++ source/backend/api/Services/ManagementFileService.cs
@@ -28,6 +28,7 @@ namespace Pims.Api.Services
         private readonly IEntityNoteRepository _entityNoteRepository;
         private readonly IManagementStatusSolver _managementStatusSolver;
         private readonly IPropertyOperationService _propertyOperationService;
+        private readonly IPropertyActivityRepository _propertyActivityRepository;
 
         public ManagementFileService(
             ClaimsPrincipal user,
@@ -40,7 +41,8 @@ namespace Pims.Api.Services
             ILookupRepository lookupRepository,
             IEntityNoteRepository entityNoteRepository,
             IManagementStatusSolver managementStatusSolver,
-            IPropertyOperationService propertyOperationService)
+            IPropertyOperationService propertyOperationService,
+            IPropertyActivityRepository propertyActivityRepository)
         {
             _user = user;
             _logger = logger;
@@ -52,6 +54,7 @@ namespace Pims.Api.Services
             _entityNoteRepository = entityNoteRepository;
             _managementStatusSolver = managementStatusSolver;
             _propertyOperationService = propertyOperationService;
+            _propertyActivityRepository = propertyActivityRepository;
         }
 
         public PimsManagementFile Add(PimsManagementFile managementFile, IEnumerable<UserOverrideCode> userOverrides)
@@ -217,13 +220,19 @@ namespace Pims.Api.Services
                 }
             }
 
+            IEnumerable<PimsPropPropActivity> fileActivityProperties = _propertyActivityRepository.GetActivitiesByManagementFile(managementFile.Internal_Id).SelectMany(pa => pa.PimsPropPropActivities);
             // The ones not on the new set should be deleted
-            List<PimsManagementFileProperty> differenceSet = currentFileProperties.Where(x => !managementFile.PimsManagementFileProperties.Any(y => y.Internal_Id == x.Internal_Id)).ToList();
+            List <PimsManagementFileProperty> differenceSet = currentFileProperties.Where(x => !managementFile.PimsManagementFileProperties.Any(y => y.Internal_Id == x.Internal_Id)).ToList();
             foreach (var deletedProperty in differenceSet)
             {
                 if (_propertyOperationService.GetOperationsForProperty(deletedProperty.PropertyId).Count > 0)
                 {
                     throw new BusinessRuleViolationException("This property cannot be deleted because it is part of a subdivision or consolidation");
+                }
+
+                if (fileActivityProperties.Any(ppa => ppa.PropertyId == deletedProperty.PropertyId))
+                {
+                    throw new BusinessRuleViolationException("This property cannot be deleted as it is part of an activity in this file");
                 }
 
                 _managementFilePropertyRepository.Delete(deletedProperty);
